Details now supports money. Also renamed variables
