+++ website/templates/product/list.html
@@ -9,6 +9,8 @@
         <ol>
         {% for item in items %}
             <li><a href="/product-categories/{{ item.id }}">{{ item.title }}</a></li>
+        {% empty %}
+            <h3> Sorry! No products</h3>
         {% endfor %}
         </ol>
     {% endif %}
@@ -18,6 +20,8 @@
         <ol>
         {% for item in items %}
             <li><a href="/products/{{ item.id }}">{{ item.title }}</a></li>
+        {% empty %}
+            <h3> Sorry! No products</h3>
         {% endfor %}
         </ol>
     {% endif %}
