+++ website/views/__init__.py
@@ -6,4 +6,4 @@ from website.views.sell_product_view import sell_product
 from website.views.user_logout_view import user_logout
 from website.views.list_product_categories_view import list_product_categories
 from website.views.product_detail_view import product_detail
-from website.views.category_detail_view import product_category_detail
+from website.views.product_category_detail_view import product_category_detail
