+++ website/views/product_category_detail_view.py
@@ -16,6 +16,10 @@ def product_category_detail(request, category_id):
 
 	# This part can probably be refactored, I just wanted to get something that works merged in
 	# (manually extracting the product data and creating a list of tuples is probably resource expensive)
+
+	
+
+	
 	new_category = [("id: ", product_category.id), ("Title: ", product_category.title)]
 	
 	return render(request, template_name, {'object_to_display': new_category, "page_title":product_category.title})
